+++ handler/user/deleteCollection.go
