+++ service/user/deleteCollection.go
@@ -1,0 +1,22 @@
+package service
+
+import (
+   "fmt"
+   "github.com/2021-ZeroGravity-backend/model"
+
+)
+
+//DeleteCollection  is used to delete collection
+    func DeleteCollection (req *model.DeleteCollectionRequest) error {
+
+	    uid := req.CollectorId;
+		id  := req.IdeaId;
+	    err := model.DeleteCollion(id, uid )
+	    if err != nil {
+		
+		  fmt.Println(err)
+	   }
+
+	 return nil 
+   
+	}
